Player can now unlock and lock doors with keys.

+++ worldmap/grid.go
@@ -95,7 +95,7 @@ func (grid *Grid) newTile(tileType string, x, y int) {
 	grid.terrain[y][x] = terrain.Icon
 	grid.passable[y][x] = terrain.Passable
 	if terrain.Door {
-		grid.door[y][x] = &doorComponent{false, rand.Int() + 1, terrain.BlocksVision, false}
+		grid.door[y][x] = &doorComponent{false, int32(rand.Int() + 1), terrain.BlocksVision, false}
 	}
 	grid.blocksVision[y][x] = terrain.BlocksVision
 }
